Split ProcessorTest
